+++ src/entities/place/api/get-seoul-data.ts
@@ -1,0 +1,5 @@
+import { seoulData } from '../model'
+
+export const getSeoulData = () => {
+  return seoulData
+}
