+++ src/entities/place/model/index.ts
@@ -1,0 +1,2 @@
+export * from './type'
+export * from './seoul-data'
