+++ src/features/course/form-course.tsx
@@ -10,7 +10,7 @@ import SelectCategories from '@/src/shared/ui/SelectCategories'
 import { RegionCascader } from '@/src/shared/ui/RegionCascader'
 import { useEffect, useState } from 'react'
 import dayjs from 'dayjs'
-import type { CoursePlanPlaceType } from '@/src/entities/place/type'
+import type { CoursePlanPlaceType } from '@/src/entities/place'
 import { Dispatch, SetStateAction } from 'react'
 import Spacer from '@/src/shared/ui/Spacer'
 import ActiveKakaoMap from '@/src/shared/ui/KakaoMap'
