+++ src/features/place/form-review.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import React, { useState, useRef, useEffect } from 'react'
 import { FieldErrors, UseFormRegister, UseFormSetValue } from 'react-hook-form'
-import { ReviewPayloadType } from '@/src/entities/place/type'
+import { ReviewPayloadType } from '@/src/entities/place'
 import { postImage } from '@/src/shared/entities/api'
 import { message } from 'antd'
 import Spacer from '@/src/shared/ui/Spacer'
@@ -12,7 +12,7 @@ interface StarRatingProps {
   setValue: UseFormSetValue<ReviewPayloadType>
 }
 
-const StarRating = ({ rating, setValue } : StarRatingProps) => {
+const StarRating = ({ rating, setValue }: StarRatingProps) => {
   const [hover, setHover] = useState(0)
 
   const handleStarClick = (ratingValue: number) => {
