+++ src/shared/ui/PlaceCollapse.tsx
@@ -5,7 +5,7 @@ import logoDefaultCopy from '@/src/assets/images/(logo)/temp_empty.png'
 import allReview from '@/src/assets/images/all_review.png'
 import kakaoReview from '@/src/assets/images/kakao_review.png'
 import StarRate from '@/src/shared/ui/StarRate'
-import type { CoursePlanPlaceType } from '@/src/entities/place/type'
+import type { CoursePlanPlaceType } from '@/src/entities/place'
 import { message } from 'antd'
 import Spacer from '@/src/shared/ui/Spacer'
 import { useState } from 'react'
