+++ src/widgets/review-comment-card/index.tsx
@@ -4,14 +4,14 @@ import ProfileImage from '@/src/shared/ui/ProfileImage'
 import Link from 'next/link'
 import { formatDateToYYYYMMDD, passFromCreate } from '@/src/shared/utils/date'
 import useUserStore from '@/src/shared/store/userStore'
-import { PlaceReviewType } from '@/src/entities/place/type'
+import { PlaceReviewType } from '@/src/entities/place'
 import emptyProfile from '@/src/assets/images/(logo)/temp_empty.png'
 import Spacer from '@/src/shared/ui/Spacer'
 import ReviewTag from '@/src/features/place/review-tag'
 import StarRate from '@/src/shared/ui/StarRate'
 import OptionDropbox from '@/src/shared/ui/OptionDropbox'
 import { useEffect, useRef, useState } from 'react'
-import { useDeletePlaceReview } from '@/src/entities/place/query'
+import { useDeletePlaceReview } from '@/src/entities/place'
 import { CommentType } from '@/src/entities/comment/type'
 import {
   useDeleteComment,
